move download file to annotations
